Рефактор
