+++ Chat/Core Layer/Networking/Pixabay/PixabayImage.swift
@@ -1,0 +1,11 @@
+//
+//  PixabayRequest.swift
+//  Chat
+//
+//  Created by Алексей Никитин on 17.11.2020.
+//  Copyright © 2020 Алексей Никитин. All rights reserved.
+//
+
+import Foundation
+
+
